calls to dict_allocate_and_serialize() are not 64-bit clean
All calls to dict_allocate_and_serialize() pass the address of a 32-bit
type, but must cast it to the 64-bit pointer type (size_t *).

This happens to work on LE machines, but even if it's apparently benign,
it's still a bug. On BE machines it is not benign.

GF_PROTOCOL_DICT_SERIALIZE() hacks around it by creating a size_t temp
var, but that's, well, a hack, IMO when you consider that all the callers
are actually passing &<u_int>; the param should just be a u_int * and
eliminate the buggy casts and the temp var in the macro.

Nobody apparently uses the Fedora/EPEL PPC RPMs, but they might. People
are trying to build gluster.org bits on SPARC and tripping over this.

BUG: 838928
Reviewed-on: http://review.gluster.com/3643
Tested-by: Gluster Build System <jenkins@build.gluster.com>

+++ libglusterfs/src/dict.c
@@ -2450,7 +2450,7 @@ out:
  */
 
 int32_t
-dict_allocate_and_serialize (dict_t *this, char **buf, size_t *length)
+dict_allocate_and_serialize (dict_t *this, char **buf, u_int *length)
 {
         int           ret    = -EINVAL;
         ssize_t       len = 0;
